+++ src/geocompy/geo/ftr.py
@@ -15,12 +15,13 @@ Types
 from __future__ import annotations
 
 from datetime import datetime
+from enum import Enum
 
 from ..data import (
     Byte,
-    parsestr,
-    parsebool,
-    toenum
+    parse_string,
+    parse_bool,
+    get_enum
 )
 from .gcdata import (
     Device,
@@ -79,11 +80,11 @@ class GeoComFTR(GeoComSubsystem):
         abort_listing
 
         """
-        _device = toenum(Device, device)
-        _filetype = toenum(File, filetype)
+        _device = get_enum(Device, device)
+        _filetype = get_enum(File, filetype)
         return self._request(
             23306,
-            [_device.value, _filetype.value, path]
+            [_device, _filetype, path]
         )
 
     def list(
@@ -150,8 +151,8 @@ class GeoComFTR(GeoComSubsystem):
             23307,
             [next],
             (
-                parsebool,
-                parsestr,
+                parse_bool,
+                parse_string,
                 int,
                 Byte.parse,
                 Byte.parse,
@@ -226,11 +227,11 @@ class GeoComFTR(GeoComSubsystem):
         abort_download
 
         """
-        _device = toenum(Device, device)
-        _filetype = toenum(File, filetype)
+        _device = get_enum(Device, device)
+        _filetype = get_enum(File, filetype)
         return self._request(
             23303,
-            [_device.value, _filetype.value, filename, blocksize],
+            [_device, _filetype, filename, blocksize],
             int
         )
 
@@ -269,7 +270,7 @@ class GeoComFTR(GeoComSubsystem):
         return self._request(
             23304,
             [block],
-            parsestr
+            parse_string
         )
 
     def abort_download(self) -> GeoComResponse[None]:
@@ -330,17 +331,17 @@ class GeoComFTR(GeoComSubsystem):
         list
 
         """
-        _device = toenum(Device, device)
-        _filetype = toenum(File, filetype)
+        _device = get_enum(Device, device)
+        _filetype = get_enum(File, filetype)
         if time is None:
-            params = [
-                _device.value, _filetype.value,
+            params: list[Enum | Byte | str] = [
+                _device, _filetype,
                 Byte(0), Byte(0), Byte(0),
                 filename
             ]
         else:
             params = [
-                _device.value, _filetype.value,
+                _device, _filetype,
                 Byte(time.day), Byte(time.month), Byte(time.year - 2000),
                 filename
             ]
@@ -389,18 +390,18 @@ class GeoComFTR(GeoComSubsystem):
         list
 
         """
-        _device = toenum(Device, device)
+        _device = get_enum(Device, device)
         _filetype = File.DATABASE
 
         if time is None:
-            params = [
-                _device.value, _filetype.value,
+            params: list[Enum | Byte | str] = [
+                _device, _filetype,
                 Byte(0), Byte(0), Byte(0),
                 dirname
             ]
         else:
             params = [
-                _device.value, _filetype.value,
+                _device, _filetype,
                 Byte(time.day), Byte(time.month), Byte(time.year - 2000),
                 dirname
             ]
@@ -456,11 +457,11 @@ class GeoComFTR(GeoComSubsystem):
         abort_download
 
         """
-        _device = toenum(Device, device)
-        _filetype = toenum(File, filetype)
+        _device = get_enum(Device, device)
+        _filetype = get_enum(File, filetype)
         return self._request(
             23313,
-            [_device.value, _filetype.value, filename, blocksize],
+            [_device, _filetype, filename, blocksize],
             int
         )
 
@@ -501,5 +502,5 @@ class GeoComFTR(GeoComSubsystem):
         return self._request(
             23314,
             [block],
-            parsestr
+            parse_string
         )