+++ CHANGELOG.md
@@ -16,10 +16,14 @@ The project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Reworked arithmetic operations supported by `Angle`
 - Reworked arithmetic operations supported by `Vector`
+- Renamed `parsestr` to `parse_string`
+- Renamed `parsebool` to `parse_bool`
+- Renamed `toenum` to `get_enum`
+- Renamed `enumparser` to `get_enum_parser`
 
 ### Fixed
 
-- GeoCom ING `set_telescopic_configuration` was missing the `prefix` parameter
+- GeoCom IMG `set_telescopic_configuration` was missing the `prefix` parameter
 - `Angle.to_dms` was sometimes returning incorrect value due to rounding
 
 ### Removed